Changing Game Over Condition
Implementing a new structure that checks the number of fainted pokemon
does one trainer has. If the trainer has no more pokemon to send out,
the game comes to an end. Right now it seems to only work for one
trainer (Anna) some brush up is necessary. The same code for Anna needs
to be implemented for the other trainer (amaya)