Fixed some battle issue + Test With squirtle.
While testing with a new pokemon i noticed that logic had some issue.
The pokemon was gaining Hp instead of loosingg HP. I quickly fixed it
with the addition of a Math.Abs.
Created a new pokemon called Squirtle he's cool. I added it to Anna's
team. I ran the earlier test with it.