More than 1 pokemon.
Anna has now 2 pokmemon in her team. Pkachu and Squirtle, when one
faints she send the next one to the field. Right now the way of calling
the second pokemon is fixed. With a bit more time and some playing
around its possible to get it so that the new pokemon is called
randomly from the list.